# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Silvercoin Loan Management API
  description: |
    Backend API for the Silvercoin loan management app. Authenticated routes use `Authorization: Bearer <accessToken>`.
    Login is email + password only (no SMS OTP).
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local (default PORT from `.env`)
  - url: http://localhost:5000
    description: Fallback if PORT unset
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with mobile and password (step 1 — sends mobile OTP)
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobile, password]
              properties:
                mobile:
                  type: string
                  description: E.164 or 10-digit India mobile
                password:
                  type: string
      responses:
        '200':
          description: OTP required before token is issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  otpRequired:
                    type: boolean
                  sessionId:
                    type: string
                  maskedMobile:
                    type: string
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedMessage'
  /api/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password (required after first login)
      operationId: authChangePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword, confirmPassword]
              properties:
                newPassword:
                  type: string
                  minLength: 6
                confirmPassword:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBody'
        '400':
          description: Validation or passwords do not match
        '401':
          $ref: '#/components/responses/UnauthorizedMessage'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/api/auth/login`
...
  responses:
    ValidationError:
      description: express-validator errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorBody'
    UnauthorizedMessage:
      description: Missing/invalid Bearer token or pending token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageBody'
...
  schemas:
    MessageBody:
      type: object
      properties:
        message:
          type: string
    ValidationErrorBody:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
...